+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from './scratchpad'
+export * from './useObservable'
+export * from './useObservableEvent'
